bot/requirements.txt
python-telegram-bot[job-queue]==20.7
requests
psutil
python-dotenv
docker
